Ensure Link.deconnect_sum completely decomposes the diagram
Previously, if the link was a diagrammatic connected sum of more than two
links, Spherogram sometimes did not return the the complete decomposition.

For example if L = L_1 # L_2 # L_3, it might return [L_1, L_2 # L_3] or
even just L itself.  (The latter seems to have only happened when there
was a pair of 2-cycles in the dual graph sharing a common edge.)